+++ common/src/main/kotlin/org/wycliffeassociates/otter/common/persistence/repositories/WorkbookRepository.kt
@@ -20,7 +20,6 @@ package org.wycliffeassociates.otter.common.persistence.repositories
 
 import com.jakewharton.rxrelay2.BehaviorRelay
 import com.jakewharton.rxrelay2.ReplayRelay
-import io.reactivex.Completable
 import io.reactivex.Maybe
 import io.reactivex.Observable
 import io.reactivex.Single
@@ -55,7 +54,6 @@ import java.util.Collections.synchronizedMap
 import javax.inject.Inject
 import org.wycliffeassociates.otter.common.persistence.IDirectoryProvider
 
-private typealias ModelTake = org.wycliffeassociates.otter.common.data.primitives.Take
 private typealias WorkbookTake = org.wycliffeassociates.otter.common.data.workbook.Take
 
 enum class PlaybackRateType {
@@ -65,7 +63,7 @@ enum class PlaybackRateType {
 
 class WorkbookRepository(
     private val directoryProvider: IDirectoryProvider,
-    private val db: IDatabaseAccessors
+    private val db: IWorkbookDatabaseAccessors
 ) : IWorkbookRepository {
     private val logger = LoggerFactory.getLogger(WorkbookRepository::class.java)
 
@@ -81,7 +79,7 @@ class WorkbookRepository(
         updateTranslationUseCase: UpdateTranslation
     ) : this(
         directoryProvider,
-        DefaultDatabaseAccessors(
+        WorkbookDatabaseAccessor(
             collectionRepository,
             contentRepository,
             resourceRepository,
@@ -617,152 +615,5 @@ class WorkbookRepository(
             }
             else -> ""
         }
-    }
-
-    interface IDatabaseAccessors {
-        fun addContentForCollection(collection: Collection, chunks: List<Content>): Completable
-        fun getChildren(collection: Collection): Single<List<Collection>>
-        fun getCollectionMetaContent(collection: Collection): Single<Content>
-        fun getContentByCollection(collection: Collection): Single<List<Content>>
-        fun getContentByCollectionActiveConnection(collection: Collection): Observable<Content>
-        fun updateContent(content: Content): Completable
-        fun getResources(content: Content, metadata: ResourceMetadata): Observable<Content>
-        fun getResources(collection: Collection, metadata: ResourceMetadata): Observable<Content>
-        fun getResourceMetadata(content: Content): List<ResourceMetadata>
-        fun getResourceMetadata(collection: Collection): List<ResourceMetadata>
-        fun getLinkedResourceMetadata(metadata: ResourceMetadata): List<ResourceMetadata>
-        fun getSubtreeResourceMetadata(collection: Collection): List<ResourceMetadata>
-        fun insertTakeForContent(take: ModelTake, content: Content): Single<Int>
-        fun getTakeByContent(content: Content): Single<List<ModelTake>>
-        fun deleteTake(take: ModelTake, date: DateHolder): Completable
-        fun getSoftDeletedTakes(metadata: ResourceMetadata, projectSlug: String): Single<List<ModelTake>>
-        fun getDerivedProject(sourceCollection: Collection): Maybe<Collection>
-        fun getDerivedProjects(): Single<List<Collection>>
-        fun getSourceProject(targetProject: Collection): Maybe<Collection>
-        fun getTranslation(sourceLanguage: Language, targetLanguage: Language): Single<Translation>
-        fun updateTranslation(translation: Translation): Completable
-        fun clearContentForCollection(
-            chapterCollection: Collection,
-            typeFilter: ContentType
-        ): Single<List<ModelTake>>
-
-        fun getChunkCount(chapterCollection: Collection): Single<Int>
-    }
-}
-
-private class DefaultDatabaseAccessors(
-    private val collectionRepo: ICollectionRepository,
-    private val contentRepo: IContentRepository,
-    private val resourceRepo: IResourceRepository,
-    private val resourceMetadataRepo: IResourceMetadataRepository,
-    private val takeRepo: ITakeRepository,
-    private val languageRepo: ILanguageRepository,
-    private val updateTranslationUseCase: UpdateTranslation
-) : WorkbookRepository.IDatabaseAccessors {
-
-    private val logger = LoggerFactory.getLogger(this::class.java)
-
-    override fun getChunkCount(chapterCollection: Collection): Single<Int> {
-        return contentRepo
-            .getByCollection(chapterCollection)
-            .map { it.count { it.type == ContentType.TEXT } }
-    }
-
-    override fun clearContentForCollection(
-        chapterCollection: Collection,
-        typeFilter: ContentType
-    ): Single<List<ModelTake>> {
-        return takeRepo
-            .getByCollection(chapterCollection, true)
-            .map {
-                it.filter { take ->
-                    takeRepo.getContentType(take).blockingGet() == typeFilter
-                }
-            }
-            .map {
-                it.forEach {
-                    takeRepo.delete(it).blockingAwait()
-                }
-                takeRepo.deleteExpiredTakes().blockingAwait()
-                contentRepo.deleteForCollection(chapterCollection, typeFilter).blockingAwait()
-                it
-            }
-    }
-
-    override fun addContentForCollection(collection: Collection, chunks: List<Content>): Completable {
-        return Observable
-            .fromArray(*chunks.toTypedArray())
-            .map { content ->
-                contentRepo.insertForCollection(content, collection)
-                    .blockingGet()
-                    .let { contentId ->
-                        content.copy(
-                            id = contentId
-                        )
-                    }
-            }
-            .toList()
-            .flatMapCompletable { contents ->
-                val sourceContents = collectionRepo.getSource(collection)
-                    .blockingGet()
-                    ?.let { collection ->
-                        contentRepo.getByCollection(collection).blockingGet()
-                    }
-
-                if (sourceContents == null || sourceContents.isEmpty()) {
-                    Completable.complete()
-                } else {
-                    contentRepo.linkDerivedToSource(contents, sourceContents)
-                }
-            }
-    }
-
-    override fun getChildren(collection: Collection) = collectionRepo.getChildren(collection)
-
-    override fun getCollectionMetaContent(collection: Collection) = contentRepo.getCollectionMetaContent(collection)
-    override fun getContentByCollection(collection: Collection) = contentRepo.getByCollection(collection)
-    override fun getContentByCollectionActiveConnection(collection: Collection): Observable<Content> {
-        return contentRepo.getByCollectionWithPersistentConnection(collection)
-    }
-
-    override fun updateContent(content: Content) = contentRepo.update(content)
-
-    override fun getResources(content: Content, metadata: ResourceMetadata) =
-        resourceRepo.getResources(content, metadata)
-
-    override fun getResources(collection: Collection, metadata: ResourceMetadata) =
-        resourceRepo.getResources(collection, metadata)
-
-    override fun getResourceMetadata(content: Content) = resourceRepo.getResourceMetadata(content)
-    override fun getResourceMetadata(collection: Collection) = resourceRepo.getResourceMetadata(collection)
-
-    override fun getLinkedResourceMetadata(metadata: ResourceMetadata) =
-        resourceMetadataRepo.getLinked(metadata).blockingGet()
-
-    override fun getSubtreeResourceMetadata(collection: Collection) =
-        resourceRepo.getSubtreeResourceMetadata(collection)
-
-    override fun insertTakeForContent(take: ModelTake, content: Content) = takeRepo.insertForContent(take, content)
-    override fun getTakeByContent(content: Content) = takeRepo.getByContent(content, includeDeleted = true)
-    override fun deleteTake(take: ModelTake, date: DateHolder) = takeRepo.update(take.copy(deleted = date.value))
-
-    override fun getSoftDeletedTakes(metadata: ResourceMetadata, projectSlug: String) =
-        takeRepo.getSoftDeletedTakes(collectionRepo.getProjectBySlugAndMetadata(projectSlug, metadata).blockingGet())
-
-    override fun getDerivedProject(sourceCollection: Collection): Maybe<Collection> {
-        return collectionRepo.getDerivedProject(sourceCollection)
-    }
-
-    override fun getDerivedProjects(): Single<List<Collection>> = collectionRepo.getDerivedProjects()
-
-    override fun getSourceProject(targetProject: Collection): Maybe<Collection> =
-        collectionRepo.getSource(targetProject)
-
-    override fun getTranslation(sourceLanguage: Language, targetLanguage: Language): Single<Translation> {
-        return languageRepo.getTranslation(sourceLanguage, targetLanguage)
-    }
-
-    override fun updateTranslation(translation: Translation): Completable {
-        return updateTranslationUseCase.update(translation)
     }
 }
