+++ jvm/workbookapp/src/main/kotlin/org/wycliffeassociates/otter/jvm/workbookapp/persistence/database/AppDatabase.kt
@@ -106,6 +106,7 @@ class AppDatabase(
     val installedEntityDao = InstalledEntityDao(dsl)
     val translationDao = TranslationDao(dsl)
     val versificationDao = VersificationDao(dsl)
+    val workbookDescriptorDao = WorkbookDescriptorDao(dsl)
 
     // Transaction support
     fun transaction(block: (DSLContext) -> Unit) {
